Create `storage::tests::vectors` module
Refactor to follow the convention used for other tests.
